+++ src/main/resources/templates/dashboard/layout/app.html
@@ -71,6 +71,7 @@
                 </div>
             </li>-->
 
+
             <li class="nav-item dropdown">
                 <a
                         class="nav-link dropdown-toggle"
@@ -79,21 +80,66 @@
                         data-toggle="dropdown"
                         aria-haspopup="true"
                         aria-expanded="false"
-                > Trần Văn Hiếu </a>
-
+                >
+                    <th:block sec:authorize="hasAuthority('ADMIN')">
+                        <span sec:authentication="principal.username"></span>
+                    </th:block>
+                    <th:block sec:authorize="hasAuthority('DOCTOR')">
+                        <span sec:authentication="principal.username"></span>
+                    </th:block>
+                    <th:block sec:authorize="isAnonymous()">
+                        <span>Tài khoản</span>
+                    </th:block>
+                    <i class="icofont-thin-down"></i
+                    ></a>
                 <ul class="dropdown-menu" aria-labelledby="dropdown02">
-
+                    <th:block sec:authorize="hasAuthority('ADMIN')">
                         <li>
-                            <a class="dropdown-item" href="/dashboard/home">Xem profile</a>
+                            <a class="dropdown-item" href="#"
+                            >Thông tin cá nhân</a
+                            >
                         </li>
-
-
                         <li>
-                            <a class="dropdown-item" href="/dashboard/logout">Đăng xuất</a>
+                            <a class="dropdown-item" href="#"
+                            >Cài đặt</a
+                            >
                         </li>
+                        <li>
+                            <a class="dropdown-item" href="/admin/logout">Đăng xuất</a>
+                        </li>
+                    </th:block>
+                    <th:block sec:authorize="hasAuthority('DOCTOR')">
+                        <li>
+                            <a class="dropdown-item" href="#"
+                            >Thông tin cá nhân</a
+                            >
+                        </li>
+                        <li>
+                            <a class="dropdown-item" href="#"
+                            >Cài đặt</a
+                            >
+                        </li>
+                        <li>
+                            <a class="dropdown-item" href="/doctor/logout">Đăng xuất</a>
+                        </li>
+                    </th:block>
+                    <th:block sec:authorize="isAnonymous()">
+                        <li>
+                            <a class="dropdown-item" href="/client/login">Đăng nhập</a>
+                        </li>
+                        <li>
+                            <a class="dropdown-item" href="/client/register">Đăng ký</a>
+                        </li>
+                    </th:block>
 
+                    <!--<th:block sec:authorize="isAuthenticated()">
+                        <li>
+                            <a class="dropdown-item" href="/doctor/logout">Đăng xuất</a>
+                        </li>
+                    </th:block>-->
                 </ul>
             </li>
+
         </ul>
     </nav>
